requirements.txt
# Core numerical libraries

numpy==1.26.4
scipy==1.11.4

# Computer vision

opencv-python-headless==4.11.0.86
Pillow==10.2.0

# Face detection & recognition

insightface==0.7.3
onnxruntime==1.17.1

faiss-cpu==1.7.4

# Machine learning utilities

scikit-learn==1.4.1.post1

# Data handling

pandas==2.2.1
tqdm==4.66.2

# API server

fastapi
uvicorn
python-multipart
python-jose
passlib
